+++ pkg/appcontext/context_test.go
@@ -1,4 +1,4 @@
-package context
+package appcontext
 
 import (
 	"context"
