+++ pkg/server/trace_test.go
@@ -6,13 +6,13 @@ import (
 
 	"github.com/google/uuid"
 
-	requestcontext "github.com/cmsgov/easi-app/pkg/context"
+	"github.com/cmsgov/easi-app/pkg/appcontext"
 )
 
 func (s ServerTestSuite) TestTraceMiddleware() {
 	// this is the actual test, since the context is cancelled post request
 	testHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		traceID, ok := requestcontext.Trace(r.Context())
+		traceID, ok := appcontext.Trace(r.Context())
 
 		s.True(ok)
 		s.NotEqual(uuid.UUID{}, traceID)
